Moved out texture/sprite caches from LOD.h/cpp
