+++ src/Engine/Graphics/ImageLoader.cpp
@@ -8,7 +8,10 @@
 #include "Engine/EngineIocContainer.h"
 #include "Engine/Graphics/IRender.h"
 #include "Engine/Graphics/Sprites.h"
+#include "Engine/Graphics/Texture_MM7.h"
 #include "Engine/Snapshots/CommonSnapshots.h"
+#include "Engine/LodTextureCache.h"
+#include "Engine/LodSpriteCache.h"
 #include "Engine/LOD.h"
 
 #include "Library/Image/ImageFunctions.h"
