+++ src/Engine/Graphics/RenderBase.cpp
@@ -8,6 +8,8 @@
 #include "Engine/SpellFxRenderer.h"
 #include "Engine/Party.h"
 #include "Engine/LOD.h"
+#include "Engine/LodTextureCache.h"
+#include "Engine/LodSpriteCache.h"
 
 #include "Engine/Objects/Actor.h"
 #include "Engine/Objects/SpriteObject.h"
