+++ src/Engine/Snapshots/CompositeSnapshots.cpp
@@ -7,6 +7,7 @@
 #include "Engine/Graphics/Level/Decoration.h"
 #include "Engine/Graphics/DecorationList.h"
 #include "Engine/Graphics/Overlays.h"
+#include "Engine/Graphics/Sprites.h"
 #include "Engine/Objects/SpriteObject.h"
 #include "Engine/Objects/ObjectList.h"
 #include "Engine/Objects/Chest.h"
